requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6  # for handling file uploads

# Environment Variables
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# External Services
google-cloud-vision>=3.4.4
openai>=1.3.0

# Testing (optional)
pytest>=7.4.3
httpx>=0.25.2  # for TestClient
pytest-asyncio>=0.21.1

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1